trabalhando nos testes de viewmodel e reposotory
