+++ ApiMercadoLivre/app/src/test/java/com/br/apimercadolivre/utils/InstantCoroutineDispatcherRule.kt
@@ -1,0 +1,36 @@
+package com.br.apimercadolivre.utils
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.TestCoroutineDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import org.junit.rules.RuleChain
+import org.junit.rules.TestWatcher
+import org.junit.runner.Description
+
+@ExperimentalCoroutinesApi
+class InstantCoroutineDispatcherRule(
+    private val testCoroutineDispatcher: TestCoroutineDispatcher = TestCoroutineDispatcher()
+) : TestWatcher() {
+
+    @ExperimentalCoroutinesApi
+    override fun starting(description: Description?) {
+        super.starting(description)
+        Dispatchers.setMain(testCoroutineDispatcher)
+    }
+
+    @ExperimentalCoroutinesApi
+    override fun finished(description: Description?) {
+        super.finished(description)
+        Dispatchers.resetMain()
+        testCoroutineDispatcher.cleanupTestCoroutines()
+    }
+
+    companion object {
+        val instantLiveDataAndCoroutineRule: RuleChain
+            get() = RuleChain.outerRule(InstantCoroutineDispatcherRule())
+                .around(InstantCoroutineDispatcherRule())
+    }
+}
+
